app/backend/requirements.txt
azure-identity==1.13.0
Flask==2.2.5
langchain==0.0.250
openai[datalib]==0.27.8
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
semantic-kernel